backend/requirements.txt
django>=5.1,<6
djangorestframework>=3.15,<4
firebase-admin>=6.6,<7
pydantic>=2.8,<3
dj-database-url>=2.2,<3
python-dotenv>=1.0,<2
psycopg2-binary>=2.9,<3
django-cors-headers>=4.4,<5
httpx>=0.27,<1
